Rework proof of problem 4 technicality in terms of countability and roots.
This accords with the goal of formalising the actual proofs I came up
with for the problems rather than a different approach chosen for
convenience of formalisation.  The result is longer than the previous
proof, but only because of the apparent lack in mathlib of results on
uncountability of reals in an interval.