Expand framework for problem 2 solution.
Easy direction of solution (that all sets as described do satisfy the
conditions of the problem) expanded in terms of lemmas that should go
in mathlib, so all "sorry"s are in proofs of those lemmas.  Harder
direction still needs to be expanded like that, and all the "sorry"s
eliminated.