Update mathlib version used.
All geometry dependencies now in mathlib.  However, there's still
euclidean_space.lean (not used in any of these solutions), where the
result on isometries should be reimplemented for mathlib in a more
general context.  And problem 2 also has a local simp lemma to work
around the lack of such lemmas for matrix notation in mathlib, which
should also be replaced by mathlib improvements.